+++ examples/cosmos-kit-lite-nextjs-pages-router-example/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.3](https://github.com/hyperweb-io/cosmos-kit/compare/example-lite@1.15.2...example-lite@1.15.3) (2025-08-12)
+
+**Note:** Version bump only for package example-lite
+
+
+
+
+
+## [1.15.2](https://github.com/hyperweb-io/cosmos-kit/compare/example-lite@1.15.1...example-lite@1.15.2) (2025-07-22)
+
+**Note:** Version bump only for package example-lite
+
+
+
+
+
+## [1.15.1](https://github.com/hyperweb-io/cosmos-kit/compare/example-lite@1.15.0...example-lite@1.15.1) (2025-07-16)
+
+**Note:** Version bump only for package example-lite
+
+
+
+
+
 # [1.15.0](https://github.com/hyperweb-io/cosmos-kit/compare/example-lite@1.14.7...example-lite@1.15.0) (2025-03-25)
 
 **Note:** Version bump only for package example-lite