+++ examples/cosmos-kit-nextjs-app-router-example/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/hyperweb-io/cosmos-kit/compare/cosmos-kit-nextjs-app-router-example@0.2.2...cosmos-kit-nextjs-app-router-example@0.2.3) (2025-08-12)
+
+**Note:** Version bump only for package cosmos-kit-nextjs-app-router-example
+
+
+
+
+
+## [0.2.2](https://github.com/hyperweb-io/cosmos-kit/compare/cosmos-kit-nextjs-app-router-example@0.2.1...cosmos-kit-nextjs-app-router-example@0.2.2) (2025-07-22)
+
+**Note:** Version bump only for package cosmos-kit-nextjs-app-router-example
+
+
+
+
+
+## [0.2.1](https://github.com/hyperweb-io/cosmos-kit/compare/cosmos-kit-nextjs-app-router-example@0.2.0...cosmos-kit-nextjs-app-router-example@0.2.1) (2025-07-16)
+
+**Note:** Version bump only for package cosmos-kit-nextjs-app-router-example
+
+
+
+
+
 # 0.2.0 (2025-03-25)
 
 **Note:** Version bump only for package cosmos-kit-nextjs-app-router-example