+++ examples/cosmos-kit-nextjs-pages-router-example/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.31.3](https://github.com/hyperweb-io/cosmos-kit/compare/example@2.31.2...example@2.31.3) (2025-08-12)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.31.2](https://github.com/hyperweb-io/cosmos-kit/compare/example@2.31.1...example@2.31.2) (2025-07-22)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.31.1](https://github.com/hyperweb-io/cosmos-kit/compare/example@2.31.0...example@2.31.1) (2025-07-16)
+
+**Note:** Version bump only for package example
+
+
+
+
+
 # [2.31.0](https://github.com/hyperweb-io/cosmos-kit/compare/example@2.30.10...example@2.31.0) (2025-03-25)
 
 **Note:** Version bump only for package example