+++ wallets/arculus-mobile/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/arculus-mobile@0.1.4...@cosmos-kit/arculus-mobile@0.1.5) (2025-08-12)
+
+**Note:** Version bump only for package @cosmos-kit/arculus-mobile
+
+
+
+
+
+## [0.1.4](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/arculus-mobile@0.1.3...@cosmos-kit/arculus-mobile@0.1.4) (2025-07-22)
+
+**Note:** Version bump only for package @cosmos-kit/arculus-mobile
+
+
+
+
+
+## [0.1.3](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/arculus-mobile@0.1.1...@cosmos-kit/arculus-mobile@0.1.3) (2025-07-16)
+
+**Note:** Version bump only for package @cosmos-kit/arculus-mobile
+
+
+
+
+
+## 0.1.2 (2025-04-29)
+
+**Note:** Change encoding to base 64, Updated deeplink and universal urls
+
 ## 0.1.1 (2025-04-07)
 
 **Note:** Version bump only for package @cosmos-kit/arculus-mobile