+++ wallets/bitgetwallet-extension/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.2](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/bitgetwallet-extension@1.2.1...@cosmos-kit/bitgetwallet-extension@1.2.2) (2025-08-12)
+
+**Note:** Version bump only for package @cosmos-kit/bitgetwallet-extension
+
+
+
+
+
+## [1.2.1](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/bitgetwallet-extension@1.2.0...@cosmos-kit/bitgetwallet-extension@1.2.1) (2025-07-22)
+
+**Note:** Version bump only for package @cosmos-kit/bitgetwallet-extension
+
+
+
+
+
+# [1.2.0](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/bitgetwallet-extension@1.1.0...@cosmos-kit/bitgetwallet-extension@1.2.0) (2025-07-16)
+
+
+### Features
+
+* update bitget wallet logo ([a6022c9](https://github.com/hyperweb-io/cosmos-kit/commit/a6022c9329ad8bfc4eb8b4943f717678a299706c))
+
+
+
+
+
 # 1.1.0 (2025-04-07)
 
 ### Features