+++ wallets/shell-extension/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.3](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/shell-extension@2.14.2...@cosmos-kit/shell-extension@2.14.3) (2025-08-12)
+
+**Note:** Version bump only for package @cosmos-kit/shell-extension
+
+
+
+
+
+## [2.14.2](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/shell-extension@2.14.1...@cosmos-kit/shell-extension@2.14.2) (2025-07-22)
+
+**Note:** Version bump only for package @cosmos-kit/shell-extension
+
+
+
+
+
+## [2.14.1](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/shell-extension@2.14.0...@cosmos-kit/shell-extension@2.14.1) (2025-07-16)
+
+**Note:** Version bump only for package @cosmos-kit/shell-extension
+
+
+
+
+
 # [2.14.0](https://github.com/hyperweb-io/cosmos-kit/compare/@cosmos-kit/shell-extension@2.13.7...@cosmos-kit/shell-extension@2.14.0) (2025-03-25)
 
 **Note:** Version bump only for package @cosmos-kit/shell-extension